make useTetrisGame simpler by adding small custom hooks
